raw and remove::wastebasket: add BD, remove archives about TipoServicos, write refactors Readme
