+++ src/logs/listen.log
@@ -26,3 +26,11 @@
 2025-05-22	01:19:11	e60a04c4-210a-4963-ad4f-3d6fbd37776f	Servidor rodando na porta 3000
 2025-05-22	01:28:18	5a9454a3-8691-45de-b01f-65a0c00f9d2d	Servidor rodando na porta 3000
 2025-05-22	14:40:19	28b166de-df7e-4b2f-968e-0cf3d3106e5c	Servidor rodando na porta 3000
+2025-07-10	09:44:54	5502c670-c933-4b12-9ebd-ff835fca1d81	Servidor rodando na porta 3000
+2025-07-10	09:47:25	9e26509d-a6af-4371-ab75-9256bfc5db19	Servidor rodando na porta 3000
+2025-07-10	09:47:56	dd16a2d0-bc71-4a03-bd37-caef52818a6a	Servidor rodando na porta 3000
+2025-07-10	09:48:15	2c9bfdba-0a38-4d47-aaac-dd7eeb3b8d41	Servidor rodando na porta 3000
+2025-07-10	09:49:04	5b52e4f7-b9a5-4001-86f9-7f116360b8d3	Servidor rodando na porta 3000
+2025-07-10	09:49:10	82d58df2-9c6f-4f67-8faf-9e9f52a0827b	Servidor rodando na porta 3000
+2025-07-10	09:51:28	1b6c3097-edba-48a6-ae29-4bbb113b64b1	Servidor rodando na porta 3000
+2025-07-10	10:36:30	ef401dbd-4454-474b-acb9-a61c15e92ce5	Servidor rodando na porta 3000
